refactor(core): promote profile apis from session routes to its own api routes
