+++ packages/core/src/routes/consts.ts
@@ -6,3 +6,6 @@ export const routes = Object.freeze({
     consent: signIn + '/consent',
   },
 });
+
+export const verificationTimeout = 10 * 60; // 10 mins.
+export const continueSignInTimeout = 10 * 60; // 10 mins.
