--- packages/core/src/routes/session/consts.ts
@@ -1,2 +1,0 @@
-export const verificationTimeout = 10 * 60; // 10 mins.
-export const continueSignInTimeout = 10 * 60; // 10 mins.
